color list and output components
