+++ src/routes/resources/rgb/index.tsx
@@ -3,15 +3,17 @@ import { routeLoader$, type DocumentHead } from '@builder.io/qwik-city';
 
 import { Gradient } from '~/components/util/HexUtils';
 import { defaults, loadPreset, v3formats, presets as presetlist } from '~/components/util/PresetUtils';
-import { convertToHex, convertToRGB, generateOutput, getBrightness, getRandomColor, getSignificantPoints, hexToHSL } from '~/components/util/RGBUtils';
+import { convertToHex, convertToRGB, disperseColors, generateOutput, getSignificantPoints, hexToHSL } from '~/components/util/RGBUtils';
 
-import { Dropdown, Toggle, NumberInput, ColorPicker } from '@luminescent/ui-qwik';
+import { Dropdown, Toggle, NumberInput } from '@luminescent/ui-qwik';
 import { inlineTranslate, useSpeak } from 'qwik-speak';
 import { getCookies, setCookies, sortColors } from '~/components/util/SharedUtils';
 import { isBrowser } from '@builder.io/qwik/build';
-import { ChevronDown, ChevronUp, Clipboard, Dices, Download, Globe, Link, Ellipsis, Palette, Save, Settings, Share, Sparkles, Trash, Type, X } from 'lucide-icons-qwik';
+import { ChevronDown, Clipboard, Download, Globe, Link, Palette, Save, Settings, Share, Sparkles, Type, X } from 'lucide-icons-qwik';
 import Input from '~/components/rgb/Input';
 import ColorMap from '~/components/rgb/ColorMap';
+import ColorList from '~/components/rgb/ColorList';
+import Output from '~/components/rgb/Output';
 
 export const rgbDefaults = {
   version: defaults.version,
@@ -54,10 +56,6 @@ export default component$(() => {
 
   const tmpstore: {
     threshold: number,
-    opened: {
-      id: number,
-      type: number,
-    },
     sectionsOpened: string[],
     alerts: {
       class: string,
@@ -65,38 +63,12 @@ export default component$(() => {
     }[],
   } = useStore({
     threshold: 50,
-    opened: {
-      id: -1,
-      type: 0,
-    },
     sectionsOpened: [],
-    previewStyle: 'default',
     alerts: [] as {
       class: string,
       text: string,
     }[],
   }, { deep: true });
-
-  const handleSwap = $((currentIndex: number, newIndex: number) => {
-    // check if the index is out of bounds
-    const colorsLength = store.colors.length;
-    if (newIndex < 0) {
-      newIndex = colorsLength - 1;
-    } else if (newIndex >= colorsLength) {
-      newIndex = 0;
-    }
-
-    const newColors = [...store.colors];
-    const currentPos = Number(`${store.colors[currentIndex].pos}`);
-    newColors[currentIndex].pos = newColors[newIndex].pos;
-    newColors[newIndex].pos = currentPos;
-    store.colors = sortColors(newColors);
-  });
-
-  const disperseColors = $(() => {
-    const newColors = store.colors.slice(0).map((color, i) => ({ hex: color.hex, pos: (100 / (store.colors.length - 1)) * i }));
-    store.colors = newColors;
-  });
 
   const decodeText = $((rgbtext: string, threshold: number) => {
     const pattern = /(?:[&§]x((?:[&§][0-9A-Fa-f]){6})|&#([0-9A-Fa-f]{6}))([^§&#]*)/;
@@ -133,7 +105,7 @@ export default component$(() => {
 
   useTask$(({ track }) => {
     if (isBrowser) setCookies('rgb', store);
-    if (store.disperse) disperseColors();
+    if (store.disperse) store.colors = disperseColors(store.colors);
     (Object.keys(store) as Array<keyof typeof store>).forEach((key) => {
       track(() => store[key]);
     });
@@ -219,143 +191,7 @@ export default component$(() => {
                 <ChevronDown size={20} />
               </div>
             </button>
-            <div class={{
-              'flex flex-col gap-2 transition-all duration-200 sm:opacity-100 sm:pointer-events-auto sm:h-auto': true,
-              'h-0 opacity-0 pointer-events-none': tmpstore.sectionsOpened.indexOf('colors') == -1,
-              'opacity-100 pointer-events-auto': tmpstore.sectionsOpened.indexOf('colors') != -1,
-            }} id="colors">
-              {store.format.color != 'MiniMessage' &&
-                <NumberInput input disabled min={1} max={store.text.length / store.colors.length} value={store.colorlength} id="colorlength" class={{ 'w-full !opacity-100': true }}
-                  onIncrement$={() => {
-                    store.colorlength++;
-                  }}
-                  onDecrement$={() => {
-                    store.colorlength--;
-                  }}
-                >
-                  {t('color.colorLength@@Characters per color')}
-                </NumberInput>
-              }
-              <NumberInput input min={2} max={store.text.length} value={store.colors.length} id="colorsinput" class={{ 'w-full': true }}
-                onChange$={(e, el) => {
-                  let colorAmount = Number(el.value);
-                  if (colorAmount < 2) return;
-                  if (colorAmount > store.text.length) return colorAmount = store.text.length;
-                  const newColors = [];
-                  for (let i = 0; i < colorAmount; i++) {
-                    if (store.colors[i]) newColors.push(store.colors[i]);
-                    else newColors.push({ hex: getRandomColor(), pos: 100 });
-                  }
-                  store.colors = newColors;
-                }}
-                onIncrement$={() => {
-                  const newColors = [...store.colors, {
-                    hex: getRandomColor(),
-                  }];
-                  store.colors = newColors.map((color, i) => ({
-                    hex: color.hex,
-                    pos: (100 / (newColors.length - 1)) * i,
-                  }));
-                }}
-                onDecrement$={() => {
-                  const newColors = store.colors.slice(0);
-                  newColors.pop();
-                  store.colors = newColors.map((color, i) => ({
-                    hex: color.hex,
-                    pos: (100 / (newColors.length - 1)) * i,
-                  }));
-                }}
-              >
-                {t('color.colorAmount@@Color Amount')}
-              </NumberInput>
-              <div class="flex gap-2">
-                <button class={{
-                  'lum-btn lum-pad-equal-xs': true,
-                  'w-full': store.disperse,
-                }} onClick$={() => {
-                  const newColors = store.colors.map(color => ({ hex: getRandomColor(), pos: color.pos }));
-                  store.colors = newColors;
-                }}>
-                  <Dices size={24} /> {store.disperse && <span>Randomize</span>}
-                </button>
-                {!store.disperse &&
-                  <button class="lum-btn lum-pad-xs w-full" disabled={store.colors.find((color, i) => color.pos != (100 / (store.colors.length - 1)) * i) ? false : true} onClick$={() => {
-                    disperseColors();
-                  }}>
-                    <Ellipsis size={24} /> Disperse
-                  </button>
-                }
-              </div>
-              <div class="flex flex-col gap-2">
-                {store.colors.map((color, i) => <div key={`${i}/${store.colors.length}`} class="flex relative gap-2">
-                  <div class="flex flex-col rounded-md">
-                    <button class="lum-btn lum-pad-equal-xs border-b-transparent rounded-b-none" onClick$={() => handleSwap(i, i - 1)}>
-                      <ChevronUp size={24} />
-                    </button>
-                    <button class="lum-btn lum-pad-equal-xs border-t-transparent rounded-t-none" onClick$={() => handleSwap(i, i + 1)}>
-                      <ChevronDown size={24} />
-                    </button>
-                  </div>
-                  <div class="flex flex-col justify-end gap-1">
-                    <label for={`colorlist-color-${i + 1}`}>{t('color.color@@Color')} {i + 1}</label>
-                    <input key={`colorlist-color-${i + 1}-${color.hex}`} id={`colorlist-color-${i + 1}`}
-                      class={{
-                        'text-gray-400 hover:text-gray-400': getBrightness(convertToRGB(color.hex)) < 126,
-                        'text-gray-700 hover:text-gray-700': getBrightness(convertToRGB(color.hex)) > 126,
-                        'lum-input w-full lum-pad-xs hover:': true,
-                      }}
-                      style={`background: ${color.hex};`}
-                      value={color.hex}
-                      onInput$={(e, el) => {
-                        const picker = document.getElementById(`colorlist-color-${i + 1}-picker`)!;
-                        picker.dataset.value = el.value;
-                        picker.dispatchEvent(new Event('input'));
-                      }}
-                      onMouseUp$={() => {
-                        const picker = document.getElementById(`colorlist-color-${i + 1}-picker`)!;
-                        picker.dataset.value = color.hex;
-                        picker.dispatchEvent(new Event('input'));
-                        if (tmpstore.opened.id == i && tmpstore.opened.type == 1) return tmpstore.opened.id = -1;
-                        else tmpstore.opened = { id: i, type: 1 };
-                        const abortController = new AbortController();
-                        document.addEventListener('click', (e) => {
-                          if (e.target instanceof HTMLElement && !e.target.closest(`#colorlist-color-${i + 1}`) && !e.target.closest(`#colorlist-color-${i + 1}-popup`)) {
-                            tmpstore.opened.id = -1;
-                            abortController.abort();
-                          }
-                        }, { signal: abortController.signal });
-                      }}
-                    />
-                  </div>
-                  <div class="flex flex-col justify-end">
-                    <button class="lum-btn lum-pad-equal-sm lum-bg-red-700 hover:lum-bg-red-600" disabled={store.colors.length <= 2} onClick$={() => {
-                      const newColors = store.colors.slice(0);
-                      newColors.splice(i, 1);
-                      store.colors = newColors;
-                    }}>
-                      <Trash size={20} />
-                    </button>
-                  </div>
-                  <div
-                    id={`colorlist-color-${i + 1}-popup`} stoppropagation:mousedown class={{
-                      'flex flex-col gap-2 motion-safe:transition-all absolute top-full z-[1000] mt-2 left-0': true,
-                      'opacity-0 scale-95 pointer-events-none': tmpstore.opened.id != i || tmpstore.opened.type != 1,
-                    }}>
-                    <ColorPicker
-                      id={`colorlist-color-${i + 1}-picker`}
-                      value={color.hex}
-                      onInput$={newColor => {
-                        const newColors = store.colors.slice(0);
-                        newColors[i].hex = newColor;
-                        store.colors = sortColors(newColors);
-                      }}
-                      showInput={false}
-                    />
-                  </div>
-                </div>,
-                )}
-              </div>
-            </div>
+            <ColorList store={store} hidden={tmpstore.sectionsOpened.indexOf('colors') == -1} />
           </div>
           <div class="flex flex-col gap-1 md:col-span-2 sm:px-2 sm:border-x border-gray-800/80" id="column2">
             <button class={{
@@ -378,56 +214,8 @@ export default component$(() => {
               </div>
             </button>
 
-            <div class={{
-              'flex flex-col gap-2 transition-all duration-200 sm:opacity-100 sm:pointer-events-auto sm:max-h-full': true,
-              'max-h-0 opacity-0 pointer-events-none': tmpstore.sectionsOpened.indexOf('output') == -1,
-              'max-h-[250px] opacity-100 pointer-events-auto': tmpstore.sectionsOpened.indexOf('output') != -1,
-            }} id="output">
-              <label for="output" class="text-gray-500">
-                {t('color.outputSubtitle@@Copy-paste this for RGB text!')}
-              </label>
-              <textarea id="output" readOnly
-                class={{
-                  'lum-input h-32 w-full font-mc whitespace-pre-wrap': true,
-                }}
-                value={generateOutput(store.text, store.colors, store.format, store.prefixsuffix, store.trimspaces, store.colorlength, store.bold, store.italic, store.underline, store.strikethrough)}
-                onClick$={(e, el) => {
-                  let alert = {
-                    class: 'text-green-500',
-                    text: 'color.copied@@Copied to clipboard!',
-                  };
-                  navigator.clipboard.writeText(el.value).catch(() => {
-                    alert = {
-                      class: 'text-red-500',
-                      text: 'color.copied@@Failed to copy to clipboard!',
-                    };
-                  });
-                  tmpstore.alerts.push(alert);
-                  setTimeout(() => {
-                    tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                  }, 2000);
-                }}
-              />
-              {tmpstore.alerts.map((alert, i) => (
-                <p key={`alert${i}`} class={alert.class} dangerouslySetInnerHTML={t(alert.text)} />
-              ))}
-              <Dropdown id="previewstyle" value={store.previewStyle} class={{ 'w-full': true }} onChange$={
-                (e, el) => {
-                  store.previewStyle = el.value;
-                }
-              } values={[
-                {
-                  name: t('color.previewstyle.default@@Default'),
-                  value: 'default',
-                },
-                {
-                  name: t('color.previewstyle.chat@@Minecraft Chat'),
-                  value: 'chat',
-                },
-              ]}>
-                {t('color.previewStyle@@Preview Style')}
-              </Dropdown>
-            </div>
+            <Output store={store} tmpstore={tmpstore} hidden={tmpstore.sectionsOpened.indexOf('output') == -1}
+              value={generateOutput(store.text, store.colors, store.format, store.prefixsuffix, store.trimspaces, store.colorlength, store.bold, store.italic, store.underline, store.strikethrough)} />
 
             <button class="lum-btn lum-bg-gray-800/30 rounded-md lum-pad-md" onClick$={() => {
               if (tmpstore.sectionsOpened.indexOf('options') == -1) tmpstore.sectionsOpened.push('options');
